# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a  Demo Template API with Swagger documentation",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3030",
    "basePath": "/pickup/v1",
    "paths": {
        "/addressdetails/": {
            "post": {
                "description": "create a new address  for a customer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addressdetails"
                ],
                "summary": "Create a new address for a customer",
                "parameters": [
                    {
                        "description": "Create address request",
                        "name": "createAddress",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createAddress"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "address created",
                        "schema": {
                            "$ref": "#/definitions/handler.AddressdetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "409": {
                        "description": "Data conflict error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
        "/addressdetails/{addressid}": {
            "put": {
                "description": "update a address's based on addressid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addressdetails"
                ],
                "summary": "Update a address",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "update address",
                        "name": "addressid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "updateAddressDetails",
                        "name": "updateAddressDetails",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateAddressDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category updated",
                        "schema": {
                            "$ref": "#/definitions/handler.AddressdetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "409": {
                        "description": "Data conflict error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a address based on addressid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addressdetails"
                ],
                "summary": "Delete a address",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "addressid",
                        "name": "addressid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Address deleted succesfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
        "/addressdetails/{customerid}": {
            "get": {
                "description": "get a address by customerid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Addressdetails"
                ],
                "summary": "Get a address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customerid ID",
                        "name": "customerid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category retrieved",
                        "schema": {
                            "$ref": "#/definitions/handler.AddressdetailsResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorValidResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.AddressdetailsResponse": {
            "type": "object",
            "properties": {
                "addressid": {
                    "type": "integer"
                },
                "addressline1": {
                    "type": "string"
                },
                "addressline2": {
                    "type": "string"
                },
                "addresstype": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "customerid": {
                    "type": "string"
                },
                "emailid": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "fromtopickup": {
                    "type": "string"
                },
                "geocode": {
                    "type": "string"
                },
                "isverified": {
                    "type": "boolean"
                },
                "landmark": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "mobilenumber": {
                    "type": "string"
                },
                "pincode": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
        "handler.createAddress": {
            "type": "object",
            "required": [
                "addressline1",
                "addressline2",
                "addresstype",
                "city",
                "country",
                "customerid",
                "emailid",
                "firstname",
                "fromtopickup",
                "geocode",
                "isverified",
                "landmark",
                "lastname",
                "mobilenumber",
                "pincode",
                "state"
            ],
            "properties": {
                "addressline1": {
                    "type": "string"
                },
                "addressline2": {
                    "type": "string"
                },
                "addresstype": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "customerid": {
                    "type": "string"
                },
                "emailid": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "fromtopickup": {
                    "type": "string"
                },
                "geocode": {
                    "type": "string"
                },
                "isverified": {
                    "type": "boolean"
                },
                "landmark": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "mobilenumber": {
                    "type": "string"
                },
                "pincode": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
...
        "handler.errorValidResponse": {
            "type": "object",
            "properties": {
                "errorno": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Error message"
                    ]
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handler.updateAddressDetails": {
            "type": "object",
            "required": [
                "addressline1",
                "addressline2",
                "addresstype",
                "city",
                "country",
                "customerid",
                "emailid",
                "firstname",
                "fromtopickup",
                "geocode",
                "isverified",
                "landmark",
                "lastname",
                "mobilenumber",
                "pincode",
                "state"
            ],
            "properties": {
                "addressline1": {
                    "type": "string",
                    "minLength": 1
                },
                "addressline2": {
                    "type": "string",
                    "minLength": 1
                },
                "addresstype": {
                    "type": "string",
                    "minLength": 1
                },
                "city": {
                    "type": "string",
                    "minLength": 1
                },
                "country": {
                    "type": "string",
                    "minLength": 1
                },
                "customerid": {
                    "type": "string",
                    "minLength": 1
                },
                "emailid": {
                    "type": "string",
                    "minLength": 1
                },
                "firstname": {
                    "type": "string",
                    "minLength": 1
                },
                "fromtopickup": {
                    "type": "string",
                    "minLength": 1
                },
                "geocode": {
                    "type": "string",
                    "minLength": 1
                },
                "isverified": {
                    "type": "boolean"
                },
                "landmark": {
                    "type": "string",
                    "minLength": 1
                },
                "lastname": {
                    "type": "string",
                    "minLength": 1
                },
                "mobilenumber": {
                    "type": "string",
                    "minLength": 1
                },
                "pincode": {
                    "type": "string",
                    "minLength": 1
                },
                "state": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
    }
}